Python/DjangoTests/Teste/requirements.txt
Django==2.2.9
djangorestframework==3.9.1
pytz==2020.1
sqlparse==0.3.1
